requirements.txt
paho-mqtt
flask
gpxpy
peewee